fix: PagedReportList 클래스의 totalElements 필드의 타입 변경

+++ src/main/java/com/example/turnpage/domain/report/converter/ReportConverter.java
@@ -60,7 +60,7 @@ public class ReportConverter {
                 .reportList(reportInfoList)
                 .page(reports.getNumber())
                 .totalPages(reports.getTotalPages())
-                .totalElements(Long.valueOf(reports.getTotalElements()).intValue())
+                .totalElements(reports.getTotalElements())
                 .isFirst(reports.isFirst())
                 .isLast(reports.isLast())
                 .build();
